Made it so more then one song wasnt playing at the same time
It was really annoying me that I could play a song then if I went to play another song they would both be playing at the same time and there was no way to stop the first song. I fixed this by making it so that anytime it plays a song it makes sure the player.stop method is ran first so that if there was a song already playing it is stopped to make way for the new song that is about to start.