Added a play random and shuffle feature
The play random feature will pick a random soung and play it once. The shuffle feature will play every song once in a random order the order will be diffrent each time you use it. NOTE: as of right now for shuffle it just plays a sample of the song because it was breaking when i tried to play the full song.